requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai==1.12.0
pydantic==2.5.3
rapidfuzz==3.6.1
cachetools==5.3.2
python-dotenv==1.0.0
httpx==0.26.0
sqlalchemy==2.0.25
pytz==2024.1
slowapi==0.1.9
gspread==5.12.0
google-auth==2.25.2

